Refactor: move authentication function into middleware folder
